session: move plugin.api.http_session
- Move `streamlink.plugin.api.http_session` module
  to `streamlink.session.http`
- Deprecate old (and unnecessary) `HTTPSession` export
  in `streamlink.plugin.api`
- Fix custom `HTTPAdapter` imports in plugins
- Fix imports and fix monkeypatch paths in tests

In the next commit:
Restore the `streamlink.plugin.api.http_session` module and re-export
`HTTPSession` and the `HTTPAdapter` classes with deprecation warnings.
