+++ src/streamlink/session/http.py
