+++ src/streamlink/session/http.pyi
